requirements.txt
blinker==1.8.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
WTForms==3.1.2
